[ci skip] patch: refactor dto -> models
