+++ backend/src/controllers/comments/get_cartoon_comments.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace dto;
+namespace models;
 
 include "../../common/response_headers.php";
 include "../../repositories/comment_repository.php";
